+++ massa-grpc/src/stream/new_transfers_info.rs
